fix: put all logic in switchButton
